Refactor and expand AMP dotcomponents picker
